Revert "migration for dispatch change"
This reverts commit f42bd69b455118abf744c875d4b948cc0a704e5d.

This is the wrong approach.  Too much effort to do the migration for too
little benefit; and Claude has demonstrated it can't put in the effort
for me.

If/when there is a database change like this, the plan will be to export
data and re-import it, rather than an online migration.